facad:run programm in terminal

+++ src/builder/product/components/GPU.java
@@ -5,7 +5,7 @@ public class GPU extends Component {
 
     private String descriptionText = "GPU: " + model + " (" + vram + " GB VRAM)";
 
-    public GPU(String model, int vram, int price) {
+    public GPU(String model, int vram, float price) {
         super(model, price);
         this.vram = vram;
     }
